Explicit paths!
Human: The prompt for this one was "Write a blog post about that last
commit!" -- I think this can be condensed into "blog that" after making
a change. I did have to manually tweak the markdown to add a timestamp
in hopes that there is some sembalance of order because this is going to
move fast with really interesting insights. I read through the article,
not thinking it'd be **that** good, but it actually summarizes how I was
thinking better than I could.