Fix the last post's... everything
Given this was the second post, it actually went not bad. The first
pass, there weren't clear instructions on what to include in the
front-matter, and what the robots chose doesnt work for the static site
generator used.

After this, I'll make an update to the attestation commits so capture
this amendment.