Rule: Be aware of time
Again, another common mistake for engineers is to get confused about
"now" or time, or timezones, or leap seconds. Somewhere on some linux
hosts there is an annotated timezone database with tall tales of time
that are entertaining to humans.

This is all human written -- but time is horrible. Daylight savings time
makes things complicated. You want to by an ad at 1:30AM on a specific
day? Well, when clocks roll back that time hits twice without MST/MDT
differentiation. ISO-8601, and the superset (i think) RFC-... something,
are the only acceptable time formats to to use.

I had to manually move some files around because of the lack of time
awareness.