Transition more of 100hellos to cortext
Human here: This was journey. Gemini-2.5 was definitely better at
tackling this problem, and how quick it is + how it breaks down the
solution is spectacular. Solving this problem was made significantly
more horrible because there seems to be a bug in Cursor's terminal when
doing something with recursive make.

This worked outside of cursor, but not inside it. That's the next thing
to tackle.