Blog post provenance
This is hitting 100% of the time now, this case was even a bit strange,
because in the last commit -- there were 2 blog articles. It was
concerned about making the change, and I pointed out wrong file -- it
looked through both files and found the one missing provenance and did
that.