Transition the site over to an AI controlled site
We demonstrate the pattern with frison/simple-sites:example, and
extended that pattern into the concept of "reflexes", which utilize the
cortex's languages and an interface bound with our nhi-entrypoint-helper
-- to enable a single-point of control across all nhi reflexes and a
unified interface.

This should be the last change demonstrating a single (and frankly, a
bit of a strange one) reflex. So with our "template" reflex --
demonstrating environment variables as input and stdout as output, the
jekyll generator demonstrating file-based IO for reflexes.