Tim here! Agentt is awesome!
Alright, so AI is trying to take credit for some things it shouldn't be
taking credit for. But I've captured the full lineage of this change and
high-level representations of our dialogue (over time). Also, it's
trying to EDIT the content to strip out my commentary which is quite
frustrating. So I'll get it committed!