AI Cleanup
AI leave a trail of garbage in it's wake -- and I haven't sorted out the
right behaviours to ensure that doesn't happen. However, asking it to do
a pass to cleanup irrelevant comments, code, and config seems to have
worked well enough!

Also, I didn't like the pattern of creating test files inline with the
tests -- so all of those were transitioned testing the concrete test
files.