Fix blog date and intro!
The date was in the future, so it wouldn't publish until tomorrow -- so
the recipe is still problematic it seems. But at least it was close?

Also, it wanted to talk about the agentt project like the whole project
is about the cli tooling (and server), but the project is about the
collaboration involved in integrating my open and closed-source agentic
tools. It just so happens this problem needed to be solved to make that
easier in the long-term.